+++ R/plotHourlyLevel.R
@@ -46,7 +46,7 @@ plotHourlyLevel <- function(x, title=NULL, units=NULL,
                             cmap=viridis_pal()(25),
                             returnData=FALSE) {
     scale <- match.arg(scale)
-    x <- checkSoundscapeInput(x, needCols='UTC')
+    x <- loadSoundscapeData(x, needCols='UTC')
     x <- toLong(x)
     if('type' %in% colnames(x)) {
         if(x$type[1] == 'BB') {